requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
compressai==1.2.2
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.33
filelock==3.9.0
fonttools==4.38.0
grad-cam==1.4.6
huggingface-hub==0.12.1
idna==3.4
importlib-resources==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.7.0.72
packaging==23.0
Pillow==9.4.0
pycocotools==2.0.6
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-msssim==0.2.1
PyYAML==6.0
requests==2.28.2
sc2bench==0.0.3
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
timm==0.8.6.dev0
torch==1.12.1
torchdistill==0.3.3
torchinfo==1.7.2
torchvision==0.13.1
tqdm==4.64.1
ttach==0.0.3
typing_extensions==4.5.0
urllib3==1.26.14
zipp==3.15.0
